Move qs rz model
