+++ wake_t/physics_models/plasma_wakefields/qs_rz_baxevanis/__init__.py
@@ -1,0 +1,3 @@
+from .wakefield import Quasistatic2DWakefield
+
+__all__ = ['Quasistatic2DWakefield']
